requirements.txt
flask==3.0.3
werkzeug==3.0.3
numpy==2.1.2
pillow==11.0.0
opencv-python-headless==4.10.0.84
transformers==4.56.1
timm==1.0.19
huggingface-hub==0.34.4
torch==2.8.0
torchvision==0.23.0
easyocr==1.7.1
